btrfs-progs: zoned: relax zone size constraint for zoned devices
Checks were in place to return error when a non power-of-2 zoned devices
is detected. Remove those checks as non power-of-2 zoned devices are
now supported.

A sane default of 1M has been chosen as the minimum alignement
requirement for zone sizes to make sure zones align with sectorsize in
different architectures.

Signed-off-by: Pankaj Raghav <p.raghav@samsung.com>